Change controllers, equals and hash with Entity, remark userdao method

+++ src/main/java/app/service/UserService.java
@@ -6,11 +6,11 @@ import java.util.List;
 public interface UserService {
     void save(User user);
 
-    void delete(int id);
+    void delete(Long id);
 
-    void update(User user, int id);
+    void update(User user, Long id);
 
     List<User> getAll();
 
-    User findById(int id);
+    User findById(Long id);
 }
